feat: move solana constant outside app directory
